Add convenience methods and fix example in readme.MD
Added some convenience methods to check the broad classification of the result (valid/invalid/error). The example in readme.MD was also not working so updated that to reflect the same concept with the new API. If you don't want to add the methods, will be happy to provide a PR with just a fixed example.